[kimchi][lookup] move impl gatetype
